requirements.txt

python-telegram-bot==13.7
gspread==3.7.0
apscheduler==3.6.3
